Move test helpers into their own folder
