+++ cli/src/tests/helpers/allocations.rs
