+++ cli/src/tests/helpers/fixtures.rs
