+++ cli/src/tests/helpers/mod.rs
@@ -1,0 +1,4 @@
+pub(super) mod allocations;
+pub(super) mod fixtures;
+pub(super) mod random;
+pub(super) mod scope_sequence;
