+++ cli/src/tests/helpers/random.rs
