+++ cli/src/tests/helpers/scope_sequence.rs
